requirements.txt
pandas
openpyxl
pytest
paramiko
pyyaml

# AI-assisted mapping (optional)
openai
